CGame: Fixed localclient rendering in splitscreen
CGame: Refactored chat slightly
CGame: Fixed weapon select for splitscreen
Particles: Tweaked emitters and adjusted textures for premultiplied alpha
ChunkWorker: Reduced visibility of baked AO
Animation: Fixed forced restart of current animation
Animation: Fixed bonecontrollers not working correctly
Weapon: Added custom projection matrix for weapons
Weapon: Started rework of weapon animation system
Weapon: New rocketlauncher and AK
Damage: Enchanced AOE dmg to check for occluders
Shader: SSAO more or less fully implemented
Shadows: Changed PCF filter to a rotated poisson disk
IO: Fixed and optimized .tga reader

+++ src/cubetech/Game/GameClient.java
@@ -886,7 +886,7 @@ public class GameClient extends Gentity implements IDieMethod, IPainMethod {
 
         ps.ducking = false;
         ps.ducked = false;
-        ps.viewheight = (int) Game.PlayerViewHeight;
+        ps.viewheight = Game.PlayerViewHeight;
         ps.ducktime = 0;
         ps.jumpTime = 0;
